backend/requirements.txt
python-multipart==0.0.6
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
segment-anything==1.0
transformers==4.35.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

